Create 01_component_2.py
I figured out the issue! This is my full code shown - component 2 was not working because it was underneath component 1. I found out that for a function like component 2 to work, it has to be at the top of the code it's applying to, or else it won't work. This may be my final commit for component 2 as it's fully working - I also edited component 1 to work with the new instructions()